backend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
environs==9.3.5
imageio==2.3.0
scikit-image==0.18.3
matplotlib==3.5.0
tqdm==4.62.3
Pillow==7.1.2
keras==2.11.0
